Started splitting up GraphService
It was 3000 lines long, it really needed to be organized. Base is probably done, but helpers need to be sorted still. This should make things easier in the future
