+++ SBOLCanvasFrontend/src/app/graph-base.spec.ts
@@ -1,0 +1,7 @@
+import { GraphBase } from './graph-base';
+
+describe('GraphBase', () => {
+  it('should create an instance', () => {
+    expect(new GraphBase()).toBeTruthy();
+  });
+});
