+++ SBOLCanvasFrontend/src/app/graph-edits.spec.ts
@@ -1,0 +1,7 @@
+import { GraphEdits } from './graph-edits';
+
+describe('GraphEdits', () => {
+  it('should create an instance', () => {
+    expect(new GraphEdits()).toBeTruthy();
+  });
+});
