+++ SBOLCanvasFrontend/src/app/graph-helpers.spec.ts
@@ -1,0 +1,7 @@
+import { GraphHelpers } from './graph-helpers';
+
+describe('GraphHelpers', () => {
+  it('should create an instance', () => {
+    expect(new GraphHelpers()).toBeTruthy();
+  });
+});
