+++ SBOLCanvasFrontend/src/app/graph.service.ts
@@ -1,8 +1,10 @@
 /*
- * GraphService
- *
- * This service controls the main editing canvas
- */
+* GraphService
+*
+* This service controls the main editing canvas
+*/
+
+/// <reference path="./graph-base.ts"/>
 
 import { Injectable } from '@angular/core';
 import * as mxEditor from 'mxgraph';
@@ -18,272 +20,28 @@ import { environment } from 'src/environments/environment';
 import { MatDialog } from '@angular/material';
 import { ErrorComponent } from './error/error.component';
 import { ConfirmComponent } from './confirm/confirm.component';
-import { mxGraphView } from 'src/mxgraph';
+import { GraphEdits } from './graph-edits';
+import { GraphBase, mx } from './graph-base';
+import { GraphHelpers } from './graph-helpers';
 
-declare var require: any;
+
+/*declare var require: any;
 const mx = require('mxgraph')({
   mxImageBasePath: 'mxgraph/images',
   mxBasePath: 'mxgraph'
-});
+});*/
 
-// Constants
-const sequenceFeatureGlyphWidth = 50;
-const sequenceFeatureGlyphHeight = 100;
-const interactionPortWidth = 10;
 
-const molecularSpeciesGlyphWidth = 50;
-const molecularSpeciesGlyphHeight = 50;
-
-const defaultTextWidth = 120;
-const defaultTextHeight = 80;
-
-const defaultInteractionSize = 80;
-
-const circuitContainerStyleName = 'circuitContainer';
-const backboneStyleName = 'backbone';
-const textboxStyleName = 'textBox';
-const scarStyleName = 'Scar (Assembly Scar)';
-const noGlyphAssignedName = 'NGA (No Glyph Assigned)';
-const molecularSpeciesGlyphBaseStyleName = 'molecularSpeciesGlyph';
-const sequenceFeatureGlyphBaseStyleName = 'sequenceFeatureGlyph';
-const interactionGlyphBaseStyleName = 'interactionGlyph';
-const moduleViewCellStyleName = "moduleViewCell";
-const componentViewCellStyleName = "componentViewCell";
-
-const interactionControlName = 'Control';
-const interactionInhibitionName = 'Inhibition';
-const interactionStimulationName = 'Stimulation';
-const interactionProcessName = 'Process';
-const interactionDegradationName = 'Degradation';
-
-// We can import non native designs. This will be set to true if ceratin annotations weren't found
-let anyForeignCellsFound = false;
 
 @Injectable({
   providedIn: 'root'
 })
-export class GraphService {
+export class GraphService extends GraphHelpers {
 
-  graph: mxGraph;
-  editor: mxEditor;
-  graphContainer: HTMLElement;
-
-  // Keeps track of the cell order we entered
-  viewStack: mxCell[];
-  selectionStack: mxCell[];
-
-  // Boolean for keeping track of whether we are showing scars or not in the graph.
-  showingScars: boolean = true;
-
-  baseMolecularSpeciesGlyphStyle: any;
-  baseSequenceFeatureGlyphStyle: any;
-
-  // This object handles the hotkeys for the graph.
-  keyHandler: any;
-
-  // edit object for interaction history
-  static interactionEdit = class {
-    cell: mxCell;
-    info: any;
-    previous: any;
-    constructor(cell, info) {
-      this.cell = cell;
-      this.info = info;
-      this.previous = info;
-    }
-
-    execute() {
-      if (this.cell != null) {
-        let tmp = this.cell.value.makeCopy();
-        if (this.previous == null) {
-          this.cell.value = null;
-        } else {
-          this.cell.value.copyDataFrom(this.previous);
-        }
-
-        this.previous = tmp;
-      }
-    }
-  }
-
-  // edit object for glyph adding history
-  static glyphInfoEdit = class {
-    cell0: mxCell;
-    info: GlyphInfo;
-    previousInfo: GlyphInfo;
-    constructor(cell0: string, info: GlyphInfo, previousInfo: GlyphInfo) {
-      this.cell0 = cell0;
-      // store them in reverse so the execute performs the action the first time
-      this.info = previousInfo;
-      this.previousInfo = info;
-    }
-
-    // Execute is called for both un-doing and re-doing
-    execute() {
-      if (this.previousInfo == null) {
-        // if the previous was null, then the dictionary didn't have an entry before so remove it
-        delete this.cell0.value[this.info.getFullURI()];
-        this.previousInfo = this.info;
-        this.info = null;
-      } else if (this.info == null) {
-        // if the current one is null, then it was removed, so re-add it
-        this.cell0.value[this.previousInfo.getFullURI()] = this.previousInfo;
-        this.info = this.previousInfo;
-        this.previousInfo = null;
-      } else {
-        // some information was changed, so update it
-        this.cell0.value[this.info.getFullURI()] = this.previousInfo;
-        const tmpInfo = this.info;
-        this.info = this.previousInfo;
-        this.previousInfo = tmpInfo;
-      }
-    }
-  }
-
-  // because the mxCurrentRootChange doesn't do what we want
-  static zoomEdit = class {
-    view: mxGraphView;
-    glyphCell: mxCell;
-    goDown: boolean;
-    graphService: GraphService;
-    constructor(view: mxGraphView, glyphCell: mxCell, graphService: GraphService) {
-      this.view = view;
-      this.glyphCell = glyphCell;
-      this.goDown = glyphCell != null;
-      this.graphService = graphService;
-    }
-
-    execute() {
-      if (this.glyphCell != null) {
-        // Zoom into the glyph
-        // get the view cell for the selected cell
-        let childViewCell;
-        if (!this.glyphCell.value) {
-          childViewCell = this.glyphCell;
-        } else {
-          childViewCell = this.graphService.graph.getModel().getCell(this.glyphCell.value);
-          // add info to the selectionstack
-          this.graphService.selectionStack.push(this.glyphCell);
-        }
-
-        // add the info to the view stack
-        this.graphService.viewStack.push(childViewCell);
-
-        // change the view
-        this.view.clear(this.view.currentRoot, true);
-        this.view.currentRoot = childViewCell;
-
-        // fix any layout problems
-        childViewCell.refreshViewCell(this.graphService.graph);
-
-        // set the selection to the circuit container
-        this.graphService.graph.setSelectionCell(childViewCell.children[0]);
-        this.graphService.fitCamera();
-
-        // make sure we can't add new strands/interactions/molecules
-        this.graphService.metadataService.setComponentDefinitionMode(this.graphService.graph.getCurrentRoot().isComponentView());
-
-        this.glyphCell = null;
-      } else {
-        // Zoom out of the glyph
-        // remove the last items on the stack
-        let previousView = this.graphService.viewStack.pop();
-        const newSelectedCell = this.graphService.selectionStack.pop();
-
-        // change the view
-        this.view.clear(this.view.currentRoot, true);
-        this.view.currentRoot = this.graphService.viewStack[this.graphService.viewStack.length - 1];
-
-        // set the selection back
-        this.graphService.graph.setSelectionCell(newSelectedCell);
-        this.graphService.setAllScars(this.graphService.showingScars);
-        this.graphService.fitCamera();
-
-        // make sure we can add new strands/interactions/molecules on the top level
-        if (this.graphService.graph.getCurrentRoot().isViewCell()) {
-          this.graphService.metadataService.setComponentDefinitionMode(this.graphService.graph.getCurrentRoot().isComponentView());
-
-          // refresh the circuit containers
-          this.graphService.graph.getCurrentRoot().refreshViewCell(this.graphService.graph);
-        }
-
-        if (newSelectedCell) {
-          this.glyphCell = newSelectedCell;
-        } else {
-          this.glyphCell = previousView
-        }
-      }
-    }
-  }
-
-  constructor(public dialog: MatDialog, private metadataService: MetadataService, private glyphService: GlyphService) {
-    // constructor code is divided into helper methods for organization,
-    // but these methods aren't entirely modular; order of some of
-    // these calls is important
-    this.initDecodeEnv();
-    this.initExtraCellMethods();
-    this.initGroupingRules();
-
-    this.graphContainer = document.createElement('div');
-    this.graphContainer.id = 'graphContainer';
-    this.graphContainer.style.margin = 'auto';
-    this.graphContainer.style.background = 'url(assets/grid.png)';
-    this.graphContainer.style.position = 'absolute';
-    this.graphContainer.style.top = '0';
-    this.graphContainer.style.bottom = '0';
-    this.graphContainer.style.left = '0';
-    this.graphContainer.style.right = '0';
-    this.graphContainer.style.overflow = 'hidden';
-
-    // mxEditor is kind of a parent to mxGraph
-    // it's used mainly for 'actions', which for now means delete, later will mean undoing
-    this.editor = new mx.mxEditor();
-    this.graph = this.editor.graph;
-    this.editor.setGraphContainer(this.graphContainer);
-
-    this.graph.setCellsCloneable(false);
-    this.graph.setConnectable(true);
-    this.graph.setDisconnectOnMove(false);
-
-    // Can't create edges without the glyph menu
-    this.graph.connectionHandler.enabled = false;
-
-    // Edges are allowed to be detached
-    this.graph.setAllowDanglingEdges(true);
-
-    // slightly clearer selection highlighting
-    mx.mxConstants.VERTEX_SELECTION_STROKEWIDTH = 2;
-
-    // Enables click-and-drag selection
-    new mx.mxRubberband(this.graph);
-
-    // This controls whether glyphs can be expanded without replacing the canvas
-    this.graph.isCellFoldable = function (cell) {
-      return false;
-      // to enable, use 'return cell.isSequenceFeatureGlyph();'
-    };
+  constructor(public dialog: MatDialog, private metadataService: MetadataService, glyphService: GlyphService) {
+    super(glyphService);
 
     this.graph.getSelectionModel().addListener(mx.mxEvent.CHANGE, mx.mxUtils.bind(this, this.handleSelectionChange));
-
-    this.initStyles();
-    this.initCustomShapes();
-    this.initSequenceFeatureGlyphMovement();
-
-    // initalize the root view cell of the graph
-    const cell1 = this.graph.getModel().getCell(1);
-    const rootViewCell = this.graph.insertVertex(cell1, "rootView", "", 0, 0, 0, 0, moduleViewCellStyleName);
-    this.graph.enterGroup(rootViewCell);
-    this.viewStack = [];
-    this.viewStack.push(rootViewCell);
-    this.selectionStack = [];
-
-    // initalize the GlyphInfoDictionary
-    const cell0 = this.graph.getModel().getCell(0);
-    const glyphDict = [];
-    this.graph.getModel().setValue(cell0, glyphDict);
-
-    // don't let any of the setup be on the undo stack
-    this.editor.undoManager.clear();
   }
 
   isRootAComponentView(): boolean {
@@ -450,10 +208,10 @@ export class GraphService {
         const geo = new mx.mxGeometry(0, 0, 0, 0);
         geo.x = 0;
         geo.y = 0;
-        geo.height = sequenceFeatureGlyphHeight;
+        geo.height = GraphBase.sequenceFeatureGlyphHeight;
 
         if (this.showingScars) {
-          geo.width = sequenceFeatureGlyphWidth;
+          geo.width = GraphBase.sequenceFeatureGlyphWidth;
         } else {
           geo.width = 0;
         }
@@ -553,7 +311,7 @@ export class GraphService {
       return;
     }
 
-    let zoomEdit = new GraphService.zoomEdit(this.graph.getView(), selection[0], this);
+    let zoomEdit = new GraphEdits.zoomEdit(this.graph.getView(), selection[0], this);
     this.graph.getModel().execute(zoomEdit);
   }
 
@@ -565,7 +323,7 @@ export class GraphService {
   exitGlyph() {
     // the root view should always be left on the viewStack
     if (this.viewStack.length > 1) {
-      let zoomEdit = new GraphService.zoomEdit(this.graph.getView(), null, this);
+      let zoomEdit = new GraphEdits.zoomEdit(this.graph.getView(), null, this);
       this.graph.getModel().execute(zoomEdit);
     }
   }
@@ -575,7 +333,7 @@ export class GraphService {
    */
   makeBackboneDragsource(element) {
     const insertGlyph = mx.mxUtils.bind(this, function (graph, evt, target, x, y) {
-      this.addBackboneAt(x - sequenceFeatureGlyphWidth / 2, y - sequenceFeatureGlyphHeight / 2);
+      this.addBackboneAt(x - GraphBase.sequenceFeatureGlyphWidth / 2, y - GraphBase.sequenceFeatureGlyphHeight / 2);
     });
 
     this.makeGeneralDragsource(element, insertGlyph);
@@ -608,12 +366,12 @@ export class GraphService {
       let glyphInfo;
       if (this.graph.getCurrentRoot().isModuleView()) {
         glyphInfo = new GlyphInfo();
-        this.addToGlyphDict(glyphInfo);
+        super.addToGlyphDict(glyphInfo);
       } else {
         glyphInfo = this.getFromGlyphDict(this.graph.getCurrentRoot().getId());
       }
-      const circuitContainer = this.graph.insertVertex(this.graph.getDefaultParent(), null, glyphInfo.getFullURI(), x, y, sequenceFeatureGlyphWidth, sequenceFeatureGlyphHeight, circuitContainerStyleName);
-      const backbone = this.graph.insertVertex(circuitContainer, null, '', 0, sequenceFeatureGlyphHeight / 2, sequenceFeatureGlyphWidth, 1, backboneStyleName);
+      const circuitContainer = this.graph.insertVertex(this.graph.getDefaultParent(), null, glyphInfo.getFullURI(), x, y, GraphBase.sequenceFeatureGlyphWidth, GraphBase.sequenceFeatureGlyphHeight, GraphBase.circuitContainerStyleName);
+      const backbone = this.graph.insertVertex(circuitContainer, null, '', 0, GraphBase.sequenceFeatureGlyphHeight / 2, GraphBase.sequenceFeatureGlyphWidth, 1, GraphBase.backboneStyleName);
 
       backbone.refreshBackbone(this.graph);
 
@@ -821,7 +579,7 @@ export class GraphService {
    */
   makeSequenceFeatureDragsource(element, stylename) {
     const insertGlyph = mx.mxUtils.bind(this, function (graph, evt, target, x, y) {
-      this.addSequenceFeatureAt(stylename, x - sequenceFeatureGlyphWidth / 2, y - sequenceFeatureGlyphHeight / 2);
+      this.addSequenceFeatureAt(stylename, x - GraphBase.sequenceFeatureGlyphWidth / 2, y - GraphBase.sequenceFeatureGlyphHeight / 2);
     });
     this.makeGeneralDragsource(element, insertGlyph);
   }
@@ -901,7 +659,7 @@ export class GraphService {
     this.graph.getModel().beginUpdate();
     try {
       // Make sure scars are/become visible if we're adding one
-      if (name.includes(scarStyleName) && !this.showingScars) {
+      if (name.includes(GraphBase.scarStyleName) && !this.showingScars) {
         this.toggleScars();
       }
 
@@ -924,15 +682,15 @@ export class GraphService {
       this.addToGlyphDict(glyphInfo);
 
       // Insert new glyph and its components
-      const sequenceFeatureCell = this.graph.insertVertex(circuitContainer, null, glyphInfo.getFullURI(), x, y, sequenceFeatureGlyphWidth, sequenceFeatureGlyphHeight, sequenceFeatureGlyphBaseStyleName + name);
+      const sequenceFeatureCell = this.graph.insertVertex(circuitContainer, null, glyphInfo.getFullURI(), x, y, GraphBase.sequenceFeatureGlyphWidth, GraphBase.sequenceFeatureGlyphHeight, GraphBase.sequenceFeatureGlyphBaseStyleName + name);
 
       // construct the view cell for it's children
       const cell1 = this.graph.getModel().getCell(1);
-      const childViewCell = this.graph.insertVertex(cell1, glyphInfo.getFullURI(), '', 0, 0, 0, 0, componentViewCellStyleName);
+      const childViewCell = this.graph.insertVertex(cell1, glyphInfo.getFullURI(), '', 0, 0, 0, 0, GraphBase.componentViewCellStyleName);
 
       // add the backbone to the child view cell
-      const childCircuitContainer = this.graph.insertVertex(childViewCell, null, glyphInfo.getFullURI(), 0, 0, 0, 0, circuitContainerStyleName);
-      const childCircuitContainerBackbone = this.graph.insertVertex(childCircuitContainer, null, '', 0, 0, 0, 0, backboneStyleName);
+      const childCircuitContainer = this.graph.insertVertex(childViewCell, null, glyphInfo.getFullURI(), 0, 0, 0, 0, GraphBase.circuitContainerStyleName);
+      const childCircuitContainerBackbone = this.graph.insertVertex(childCircuitContainer, null, '', 0, 0, 0, 0, GraphBase.backboneStyleName);
 
       childCircuitContainerBackbone.setConnectable(false);
       childCircuitContainer.setConnectable(false);
@@ -1019,7 +777,7 @@ export class GraphService {
    */
   makeMolecularSpeciesDragsource(element, stylename) {
     const insertGlyph = mx.mxUtils.bind(this, function (graph, evt, target, x, y) {
-      this.addMolecularSpeciesAt(stylename, x - molecularSpeciesGlyphWidth / 2, y - molecularSpeciesGlyphHeight / 2);
+      this.addMolecularSpeciesAt(stylename, x - GraphBase.molecularSpeciesGlyphWidth / 2, y - GraphBase.molecularSpeciesGlyphHeight / 2);
     });
     this.makeGeneralDragsource(element, insertGlyph);
   }
@@ -1066,7 +824,7 @@ export class GraphService {
       this.addToGlyphDict(proteinInfo);
 
       const molecularSpeciesGlyph = this.graph.insertVertex(this.graph.getDefaultParent(), null, proteinInfo.getFullURI(), x, y,
-        molecularSpeciesGlyphWidth, molecularSpeciesGlyphHeight, molecularSpeciesGlyphBaseStyleName + name);
+        GraphBase.molecularSpeciesGlyphWidth, GraphBase.molecularSpeciesGlyphHeight, GraphBase.molecularSpeciesGlyphBaseStyleName + name);
       molecularSpeciesGlyph.setConnectable(true);
 
       // The new glyph should be selected
@@ -1084,7 +842,7 @@ export class GraphService {
    */
   makeInteractionDragsource(element, stylename) {
     const insertGlyph = mx.mxUtils.bind(this, function (graph, evt, target, x, y) {
-      this.addInteractionAt(stylename, x - defaultInteractionSize / 2, y - defaultInteractionSize / 2);
+      this.addInteractionAt(stylename, x - GraphBase.defaultInteractionSize / 2, y - GraphBase.defaultInteractionSize / 2);
     });
     this.makeGeneralDragsource(element, insertGlyph);
   }
@@ -1124,12 +882,12 @@ export class GraphService {
 
     this.graph.getModel().beginUpdate();
     try {
-      cell = new mx.mxCell(new InteractionInfo(), new mx.mxGeometry(x, y, 0, 0), interactionGlyphBaseStyleName + name);
+      cell = new mx.mxCell(new InteractionInfo(), new mx.mxGeometry(x, y, 0, 0), GraphBase.interactionGlyphBaseStyleName + name);
 
       const selectionCells = this.graph.getSelectionCells();
       if (selectionCells.length == 1) {
         const selectedCell = this.graph.getSelectionCell();
-        cell.geometry.setTerminalPoint(new mx.mxPoint(x, y - defaultInteractionSize), false);
+        cell.geometry.setTerminalPoint(new mx.mxPoint(x, y - GraphBase.defaultInteractionSize), false);
         cell.edge = true;
         this.graph.addEdge(cell, null, selectedCell, null);
       } else if (selectionCells.length == 2) {
@@ -1138,8 +896,8 @@ export class GraphService {
         cell.edge = true;
         this.graph.addEdge(cell, null, sourceCell, destCell);
       } else {
-        cell.geometry.setTerminalPoint(new mx.mxPoint(x, y + defaultInteractionSize), true);
-        cell.geometry.setTerminalPoint(new mx.mxPoint(x + defaultInteractionSize, y), false);
+        cell.geometry.setTerminalPoint(new mx.mxPoint(x, y + GraphBase.defaultInteractionSize), true);
+        cell.geometry.setTerminalPoint(new mx.mxPoint(x + GraphBase.defaultInteractionSize, y), false);
         cell.edge = true;
         this.graph.addEdge(cell, null, null, null);
       }
@@ -1183,14 +941,14 @@ export class GraphService {
     try {
       graph.getModel().beginUpdate();
       // make sure the glyph style matches the partRole
-      let newStyleName = sequenceFeatureGlyphBaseStyleName + name;
+      let newStyleName = GraphBase.sequenceFeatureGlyphBaseStyleName + name;
 
       // if there's no style for the partRole, use noGlyphAssigned
       let cellStyle = graph.getStylesheet().getCellStyle(newStyleName);
       // if there is no registered style for the newStyleName, getCellStyle returns an empty object.
       // all of our registered styles have several fields, use fillcolor as an example to check
       if (!cellStyle.fillColor)
-        newStyleName = sequenceFeatureGlyphBaseStyleName + noGlyphAssignedName;
+        newStyleName = GraphBase.sequenceFeatureGlyphBaseStyleName + GraphBase.noGlyphAssignedName;
 
       cells.forEach(function (cell) {
         if (cell.isSequenceFeatureGlyph()) {
@@ -1202,7 +960,7 @@ export class GraphService {
           } else {
             // the string is something like "strokecolor=#000000;glyphStyleName;fillcolor=#ffffff;etc;etc;"
             // we only want to replace the 'glyphStyleName' bit
-            let startIdx = styleString.indexOf(sequenceFeatureGlyphBaseStyleName);
+            let startIdx = styleString.indexOf(GraphBase.sequenceFeatureGlyphBaseStyleName);
             let endIdx = styleString.indexOf(';', startIdx);
             let stringToReplace = styleString.slice(startIdx, endIdx - startIdx);
             styleString = styleString.replace(stringToReplace, newStyleName);
@@ -1233,7 +991,7 @@ export class GraphService {
         if (name == "Biochemical Reaction" || name == "Non-Covalent Binding" || name == "Genetic Production") {
           name = "Process";
         }
-        name = interactionGlyphBaseStyleName + name;
+        name = GraphBase.interactionGlyphBaseStyleName + name;
 
         // Modify the style string
         let styleString = selectedCell.style.slice();
@@ -1243,7 +1001,7 @@ export class GraphService {
         } else {
           // the string is something like "strokecolor=#000000;interactionStyleName;fillcolor=#ffffff;etc;etc;"
           // we only want to replace the 'glyphStyleName' bit
-          let startIdx = styleString.indexOf(interactionGlyphBaseStyleName);
+          let startIdx = styleString.indexOf(GraphBase.interactionGlyphBaseStyleName);
           let endIdx = styleString.indexOf(';', startIdx);
           let stringToReplace = styleString.slice(startIdx, endIdx - startIdx);
           styleString = styleString.replace(stringToReplace, name);
@@ -1292,7 +1050,7 @@ export class GraphService {
    */
   makeTextboxDragsource(element) {
     const insert = mx.mxUtils.bind(this, function (graph, evt, target, x, y) {
-      this.addTextBoxAt(x - defaultTextWidth / 2, y - defaultTextHeight / 2);
+      this.addTextBoxAt(x - GraphBase.defaultTextWidth / 2, y - GraphBase.defaultTextHeight / 2);
     });
     this.makeGeneralDragsource(element, insert);
   }
@@ -1311,7 +1069,7 @@ export class GraphService {
   addTextBoxAt(x, y) {
     this.graph.getModel().beginUpdate();
     try {
-      const cell = this.graph.insertVertex(this.graph.getDefaultParent(), null, 'Sample Text', x, y, defaultTextWidth, defaultTextHeight, textboxStyleName);
+      const cell = this.graph.insertVertex(this.graph.getDefaultParent(), null, 'Sample Text', x, y, GraphBase.defaultTextWidth, GraphBase.defaultTextHeight, GraphBase.textboxStyleName);
       cell.setConnectable(false);
 
       // The new cell should be selected
@@ -1320,39 +1078,6 @@ export class GraphService {
     } finally {
       this.graph.getModel().endUpdate();
     }
-  }
-
-  /**
-   * Updates a GlyphInfo
-   * NOTE: Should only be used if the glyphURI is the same
-   */
-  private updateGlyphDict(info: GlyphInfo) {
-    const cell0 = this.graph.getModel().getCell(0);
-    this.graph.getModel().execute(new GraphService.glyphInfoEdit(cell0, info, cell0.value[info.getFullURI()]));
-  }
-
-  /**
-   * Remove a GlyphInfo object from the dictionary
-   */
-  private removeFromGlyphDict(glyphURI: string) {
-    const cell0 = this.graph.getModel().getCell(0);
-    this.graph.getModel().execute(new GraphService.glyphInfoEdit(cell0, null, cell0.value[glyphURI]));
-  }
-
-  /**
-   * Add a GlyphInfo object to the dictionary
-   */
-  private addToGlyphDict(info: GlyphInfo) {
-    const cell0 = this.graph.getModel().getCell(0);
-    this.graph.getModel().execute(new GraphService.glyphInfoEdit(cell0, info, null));
-  }
-
-  /**
-   * Get the GlyphInfo with the given glyphURI from the dictionary
-   */
-  getFromGlyphDict(glyphURI: string): GlyphInfo {
-    const cell0 = this.graph.getModel().getCell(0);
-    return cell0.value[glyphURI];
   }
 
   /**
@@ -1413,7 +1138,7 @@ export class GraphService {
               if (this.graph.getCurrentRoot().isComponentView()) {
                 // default case (we're zoomed into a component view)
                 const glyphZoomed = this.selectionStack[this.selectionStack.length - 1];
-                this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), null, this));
+                this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), null, this));
 
                 // default case
                 // update the viewcell ID
@@ -1423,7 +1148,7 @@ export class GraphService {
                 // update the selected cell id
                 this.graph.getModel().setValue(glyphZoomed, newGlyphURI);
 
-                this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), glyphZoomed, this));
+                this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), glyphZoomed, this));
               } else {
                 // edge case (we're in a module view, and there is no view cell to update)
                 this.graph.getModel().setValue(selectedCell, newGlyphURI);
@@ -1464,7 +1189,7 @@ export class GraphService {
           this.updateAngularMetadata(this.graph.getSelectionCells());
 
         } else {
-          let interactionEdit = new GraphService.interactionEdit(selectedCell, info);
+          let interactionEdit = new GraphEdits.interactionEdit(selectedCell, info);
           this.mutateInteractionGlyph(info.interactionType);
           this.graph.getModel().execute(interactionEdit);
         }
@@ -1505,7 +1230,7 @@ export class GraphService {
           if (selectedCells.length == 1 && selectedCells[0].isCircuitContainer()) {
             const glyphZoomed = this.selectionStack[this.selectionStack.length - 1];
             // zoom out before making changes
-            this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), null, this));
+            this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), null, this));
 
             // remove the found viewCell and replace it with this one
             this.removeViewCell(conflictViewCell);
@@ -1517,7 +1242,7 @@ export class GraphService {
             this.updateViewCell(viewCell, newGlyphURI);
 
             // rezoom after the ID has been changed
-            this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), glyphZoomed, this));
+            this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), glyphZoomed, this));
           } else {
             // remove the found viewCell and replace it with this one
             this.removeViewCell(conflictViewCell);
@@ -1544,14 +1269,14 @@ export class GraphService {
 
             const glyphZoomed = this.selectionStack[this.selectionStack.length - 1];
             // unzoom and rezoom after the rename
-            this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), null, this));
+            this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), null, this));
 
             // remove this cells viewCell and update the glyphURI
             const viewCell = this.graph.getModel().getCell(oldGlyphURI);
             this.removeViewCell(viewCell);
 
             this.graph.getModel().setValue(glyphZoomed, newGlyphURI);
-            this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), glyphZoomed, this));
+            this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), glyphZoomed, this));
           } else {
             // remove this cells viewCell and update the glyphURI
             const viewCell = this.graph.getModel().getCell(oldGlyphURI);
@@ -1598,7 +1323,7 @@ export class GraphService {
         // if the current selectedcell is a circuit container we need to unzoom and rezoom
         if (selectedCell.isCircuitContainer()) {
           const zoomedCell = this.selectionStack[this.selectionStack.length - 1];
-          this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), null, this));
+          this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), null, this));
           this.updateViewCell(viewCell, newGlyphURI);
 
           // update the glyphURI of the other cells
@@ -1608,7 +1333,7 @@ export class GraphService {
             graph.getModel().setValue(cell, newGlyphURI);
           });
 
-          this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), zoomedCell, this));
+          this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), zoomedCell, this));
         } else {
           this.updateViewCell(viewCell, newGlyphURI);
 
@@ -1636,9 +1361,9 @@ export class GraphService {
 
         if (selectedCell.isCircuitContainer()) {
           const glyphZoomed = this.selectionStack[this.selectionStack.length - 1];
-          this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), null, this));
+          this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), null, this));
           this.graph.getModel().setValue(glyphZoomed, newGlyphURI);
-          this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), glyphZoomed, this));
+          this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), glyphZoomed, this));
         } else {
           // update the selected cell's glyphURI
           this.graph.getModel().setValue(selectedCell, newGlyphURI);
@@ -1877,7 +1602,7 @@ export class GraphService {
       // more than the direct parent's view cell may have been changed, so it's easiest to zoom all the way out (trust me, It had problems when it wasn't zooming out)
       let zoomedCells = this.selectionStack.slice();
       for (let i = 0; i < zoomedCells.length; i++) {
-        this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), null, this));
+        this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), null, this));
       }
 
       // zoom out of the rootView TODO come back to me when recursive modules is a thing
@@ -1887,7 +1612,7 @@ export class GraphService {
         rootViewId = this.graph.getCurrentRoot().getId();
         rootViewInfo = this.getFromGlyphDict(rootViewId);
         if (rootViewInfo) {
-          this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), null, this));
+          this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), null, this));
         }
       } else {
         // special case of the root getting changed before changeOwnership is called
@@ -1939,11 +1664,11 @@ export class GraphService {
         rootViewInfo = rootViewInfo.makeCopy();
         rootViewInfo.uriPrefix = GlyphInfo.baseURI;
         let newRootViewCell = this.graph.getModel().getCell(rootViewInfo.getFullURI());
-        this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), newRootViewCell, this));
+        this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), newRootViewCell, this));
       }
       // re zoom to fix the view
       for (let i = 0; i < zoomedCells.length; i++) {
-        this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), zoomedCells[i], this));
+        this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), zoomedCells[i], this));
       }
     } finally {
       this.graph.getModel().endUpdate();
@@ -2051,7 +1776,7 @@ export class GraphService {
    * and uses it to replace the current graph
    */
   setGraphToXML(graphString: string) {
-    anyForeignCellsFound = false;
+    this.anyForeignCellsFound = false;
     this.graph.home();
     this.graph.getModel().clear();
 
@@ -2080,10 +1805,10 @@ export class GraphService {
         element.refreshCircuitContainer(this.graph);
     });
 
-    if (anyForeignCellsFound) {
+    if (this.anyForeignCellsFound) {
       console.log("FORMATTING !!!!!!!!!!!!!!!!");
       this.autoFormat();
-      anyForeignCellsFound = false;
+      this.anyForeignCellsFound = false;
     }
 
     this.fitCamera();
@@ -2131,7 +1856,7 @@ export class GraphService {
         // zoom out to make things easier
         if (fromCircuitContainer && this.viewStack.length > 1) {
           selectedCell = this.selectionStack[this.selectionStack.length - 1];
-          this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), null, this));
+          this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), null, this));
         }
 
         // setup the decoding info
@@ -2219,7 +1944,7 @@ export class GraphService {
         if (fromCircuitContainer) {
           if (selectedCell.isSequenceFeatureGlyph()) {
             // if the selected cell is a sequenceFeature that means we came from a sub view
-            this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), newCell, this));
+            this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), newCell, this));
           } else {
             // if it isn't, that means we are in a module, or at the root
             let newRootView = this.graph.getModel().getCell(newCell.getValue());
@@ -2236,7 +1961,7 @@ export class GraphService {
               //circuitContainer.setCollapsed(false); 
             } else {
               // at the root, just zoom back in
-              this.graph.getModel().execute(new GraphService.zoomEdit(this.graph.getView(), newRootView, this));
+              this.graph.getModel().execute(new GraphEdits.zoomEdit(this.graph.getView(), newRootView, this));
               circuitContainer.refreshCircuitContainer(this.graph);
             }
           }
@@ -2265,13 +1990,13 @@ export class GraphService {
 
     // initalize the root view cell of the graph
     if (moduleMode) {
-      rootViewCell = this.graph.insertVertex(cell1, "rootView", "", 0, 0, 0, 0, moduleViewCellStyleName);
+      rootViewCell = this.graph.insertVertex(cell1, "rootView", "", 0, 0, 0, 0, GraphBase.moduleViewCellStyleName);
       this.graph.enterGroup(rootViewCell);
       this.viewStack.push(rootViewCell);
     } else {
       let info = new GlyphInfo();
       this.addToGlyphDict(info);
-      rootViewCell = this.graph.insertVertex(cell1, info.getFullURI(), "", 0, 0, 0, 0, componentViewCellStyleName);
+      rootViewCell = this.graph.insertVertex(cell1, info.getFullURI(), "", 0, 0, 0, 0, GraphBase.componentViewCellStyleName);
       this.graph.enterGroup(rootViewCell);
       this.viewStack.push(rootViewCell);
       this.addBackbone();
@@ -2283,775 +2008,15 @@ export class GraphService {
   }
 
   /**
-   * Sets up environment variables to make decoding new graph models from xml into memory
+   * Sets the graph to component definition mode or module mode.
+   * wrapper for metadataService.setComponentDefinitionMode.
+   * @param componentMode True if you want to be in component mode, false if module mode.
    */
-  initDecodeEnv() {
-    // stuff needed for decoding
-    window['mxGraphModel'] = mx.mxGraphModel;
-    window['mxGeometry'] = mx.mxGeometry;
-    window['mxPoint'] = mx.mxPoint;
-    window['mxCell'] = mx.mxCell;
-
-    //mxGraph uses function.name which uglifyJS breaks on production
-    // Glyph info decode/encode
-    Object.defineProperty(GlyphInfo, "name", { configurable: true, value: "GlyphInfo" });
-    const glyphInfoCodec = new mx.mxObjectCodec(new GlyphInfo());
-    glyphInfoCodec.decode = function (dec, node, into) {
-      const glyphData = new GlyphInfo();
-      const meta = node;
-      if (meta != null) {
-        for (let i = 0; i < meta.attributes.length; i++) {
-          const attrib = meta.attributes[i];
-          if (attrib.specified == true && attrib.name != 'as') {
-            glyphData[attrib.name] = attrib.value;
-          }
-        }
-        for (let i = 0; i < meta.children.length; i++) {
-          const childNode = meta.children[i];
-          if (childNode.getAttribute("as") === "otherTypes") {
-            glyphData.otherTypes = dec.decode(childNode);
-          } else if (childNode.getAttribute("as") === "otherRoles") {
-            glyphData.otherRoles = dec.decode(childNode);
-          }
-        }
-      }
-      return glyphData;
-    }
-    glyphInfoCodec.encode = function (enc, object) {
-      return object.encode(enc);
-    }
-    mx.mxCodecRegistry.register(glyphInfoCodec);
-    window['GlyphInfo'] = GlyphInfo;
-
-    // Interaction info encode/decode
-    Object.defineProperty(InteractionInfo, "name", { configurable: true, value: "InteractionInfo" });
-    const interactionInfoCodec = new mx.mxObjectCodec(new InteractionInfo());
-    interactionInfoCodec.decode = function (dec, node, into) {
-      const interactionData = new InteractionInfo();
-      const meta = node;
-      if (meta != null) {
-        for (let i = 0; i < meta.attributes.length; i++) {
-          const attrib = meta.attributes[i];
-          if (attrib.specified == true && attrib.name != 'as') {
-            interactionData[attrib.name] = attrib.value;
-          }
-        }
-      }
-      return interactionData;
-    }
-    interactionInfoCodec.encode = function (enc, object) {
-      return object.encode(enc);
-    }
-    mx.mxCodecRegistry.register(interactionInfoCodec);
-    window['InteractionInfo'] = InteractionInfo;
-
-    // For circuitContainers, the order of the children matters.
-    // We want it to match the order of the children's geometries
-    const defaultDecodeCell = mx.mxCodec.prototype.decodeCell;
-    mx.mxCodec.prototype.decodeCell = function (node, restoreStructures) {
-      const cell = defaultDecodeCell.apply(this, arguments);
-
-      // find cell 0 for the glyph dict
-      let cell0 = cell;
-      while (cell0.getId() != "0") {
-        cell0 = cell0.parent;
-      }
-      let glyphDict = cell0.value;
-
-      // check for format conditions
-      if ((cell.isCircuitContainer() && cell.getParent().getId() === "rootView" || cell.isMolecularSpeciesGlyph()) && cell.getGeometry().height == 0) {
-        anyForeignCellsFound = true;
-      }
-
-      // reconstruct the cell style
-      if (cell && cell.id > 1 && (cell.style == null || cell.style.length == 0 || (!cell.isViewCell() && cell.getGeometry().height == 0))) {
-        if (glyphDict[cell.value] != null) {
-          if (glyphDict[cell.value].partType === 'DNA region') {
-            // sequence feature
-            if (!cell.style) {
-              cell.style = sequenceFeatureGlyphBaseStyleName + glyphDict[cell.value].partRole;
-            } else {
-              cell.style = cell.style.replace(sequenceFeatureGlyphBaseStyleName, sequenceFeatureGlyphBaseStyleName + glyphDict[cell.value].partRole);
-            }
-            cell.geometry.width = sequenceFeatureGlyphWidth;
-            cell.geometry.height = sequenceFeatureGlyphHeight;
-          } else {
-            // molecular species
-            cell.style = molecularSpeciesGlyphBaseStyleName + "macromolecule";
-            cell.geometry.width = molecularSpeciesGlyphWidth;
-            cell.geometry.height = molecularSpeciesGlyphHeight;
-          }
-        } else if (cell.value instanceof InteractionInfo) {
-          // interaction
-          let name = cell.value.interactionType;
-          if (name == "Biochemical Reaction" || name == "Non-Covalent Binding" || name == "Genetic Production") {
-            name = "Process";
-          }
-          cell.style = interactionGlyphBaseStyleName + name;
-        }
-      }
-
-      if (cell && cell.isSequenceFeatureGlyph()) {
-        cell.parent.children.sort(function (cellA, cellB) {
-          return cellA.getGeometry().x - cellB.getGeometry().x;
-        });
-      }
-      return cell;
-    }
+  public setComponentDefinitionMode(componentMode: boolean) {
+    this.metadataService.setComponentDefinitionMode(componentMode);
   }
 
-  /**
-   * Gives mxCells new methods related to our circuit/backbone rules
-   */
-  initExtraCellMethods() {
 
-    mx.mxCell.prototype.isStyle = function (styleName) {
-      if (!this.style)
-        return false;
-      return this.style.includes(styleName);
-    };
-
-    mx.mxCell.prototype.isBackbone = function () {
-      return this.isStyle(backboneStyleName);
-    };
-
-    mx.mxCell.prototype.isMolecularSpeciesGlyph = function () {
-      return this.isStyle(molecularSpeciesGlyphBaseStyleName);
-    };
-
-    mx.mxCell.prototype.isCircuitContainer = function () {
-      return this.isStyle(circuitContainerStyleName);
-    };
-
-    mx.mxCell.prototype.isSequenceFeatureGlyph = function () {
-      return this.isStyle(sequenceFeatureGlyphBaseStyleName);
-    };
-
-    mx.mxCell.prototype.isScar = function () {
-      return this.isStyle(scarStyleName);
-    };
-
-    mx.mxCell.prototype.isInteraction = function () {
-      return this.isStyle(interactionGlyphBaseStyleName);
-    }
-
-    mx.mxCell.prototype.isViewCell = function () {
-      return this.isStyle(moduleViewCellStyleName) || this.isStyle(componentViewCellStyleName);
-    }
-
-    mx.mxCell.prototype.isModuleView = function () {
-      return this.isStyle(moduleViewCellStyleName);
-    }
-
-    mx.mxCell.prototype.isComponentView = function () {
-      return this.isStyle(componentViewCellStyleName);
-    }
-
-    /**
-     * Returns the id of the cell's highest ancestor
-     * (or the cell's own id if it has no parent)
-     */
-    mx.mxCell.prototype.getRootId = function () {
-      if (this.isSequenceFeatureGlyph()) {
-        return this.parent.getId();
-      } else {
-        return this.getId();
-      }
-    }
-
-    /**
-     * Returns the backbone associated with this cell
-     */
-    mx.mxCell.prototype.getBackbone = function () {
-      if (this.isSequenceFeatureGlyph()) {
-        return this.getCircuitContainer().getBackbone();
-      } else if (this.isBackbone()) {
-        return this;
-      } else if (!this.isCircuitContainer()) {
-        return null;
-      }
-
-      for (let i = 0; i < this.children.length; i++) {
-        if (this.children[i].isBackbone()) {
-          return this.children[i]
-        }
-      }
-
-      console.error("getBackbone(): No backbone found in circuit container!");
-      return null;
-    };
-
-    mx.mxCell.prototype.getSequenceFeatureGlyph = function () {
-      if (this.isSequenceFeatureGlyph()) {
-        return this;
-      }
-      else if (this.isBackbone()) {
-        return this.getParent().getSequenceFeatureGlyph();
-      }
-      else if (this.isCircuitContainer()) {
-        if (this.getParent().isSequenceFeatureGlyph()) {
-          return this.getParent();
-        } else {
-          // top level circuitContainers have no containing sequenceFeatureGlyph
-          return null;
-        }
-      }
-
-      return null;
-    };
-
-    /**
-     * Positions and sizes the backbone associated with this cell
-     */
-    mx.mxCell.prototype.refreshBackbone = function (graph) {
-      if (this.isBackbone()) {
-        this.getCircuitContainer(graph).refreshBackbone(graph);
-        return;
-      } else if (!this.isCircuitContainer()) {
-        console.error("refreshBackbone: called on an invalid cell!");
-        return;
-      }
-      // NOTE: 'this' is the circuitContainer, not the backbone
-      // (for easier access to the list of glyphs)
-
-      // width:
-      let width = 0;
-      for (let i = 0; i < this.children.length; i++) {
-        if (this.children[i].isSequenceFeatureGlyph()) {
-          width += this.children[i].getGeometry().width;
-        }
-      }
-      if (width < sequenceFeatureGlyphWidth) {
-        width = sequenceFeatureGlyphWidth;
-      }
-
-      // Shape is a line, not rectangle, so any non-zero height is fine
-      let height = 1;
-
-      this.getBackbone().replaceGeometry('auto', sequenceFeatureGlyphHeight / 2, width, height, graph);
-    };
-
-    /**
-     * (Re)positions the glyphs inside the circuitContainer and
-     * also refreshes the backbone.
-     */
-    mx.mxCell.prototype.refreshCircuitContainer = function (graph) {
-      if (!this.isCircuitContainer()) {
-        console.error("refreshCircuitContainer: called on an invalid cell!");
-        return;
-      }
-
-      // refresh backbone (width, height)
-      this.refreshBackbone(graph);
-
-      // verify own width, height
-      this.replaceGeometry(
-        'auto', 'auto', this.getBackbone().getGeometry().width, sequenceFeatureGlyphHeight, graph);
-
-      // put the backbone first in the children array so it is drawn before glyphs
-      // (meaning it appears behind them)
-      if (!this.children[0].isBackbone())
-        graph.getModel().add(this, this.getBackbone(), 0);
-
-      // Layout all the glyphs in a horizontal line, while ignoring the backbone cell.
-      const layout = new mx.mxStackLayout(graph, true);
-      layout.resizeParent = true;
-      layout.isVertexIgnored = function (vertex) {
-        return vertex.isBackbone()
-      };
-      layout.execute(this);
-    };
-
-    mx.mxCell.prototype.refreshViewCell = function (graph) {
-      if (!this.isViewCell()) {
-        console.error("refreshViewCell: called on an invalid cell!");
-        return;
-      }
-
-      // refresh all circuit containers
-      for (let child of this.children) {
-        if (child.isCircuitContainer()) {
-          child.refreshCircuitContainer(graph);
-        }
-      }
-    };
-
-    /**
-     * Returns the circuit container associated with this cell.
-     */
-    mx.mxCell.prototype.getCircuitContainer = function (graph) {
-      if (this.isSequenceFeatureGlyph()) {
-        const children = graph.getModel().getCell(this.value).children;
-        for (let child of children) {
-          if (child.isCircuitContainer()) {
-            return child;
-          }
-        }
-      } else if (this.isCircuitContainer()) {
-        return this;
-      } else if (this.isBackbone()) {
-        return this.getParent();
-      }
-
-      return null;
-    };
-
-    /**
-     * Replaces this cell's geometry in an undo friendly way
-     * 'graph' must be a reference to the graph
-     *
-     * For any other value, pass the string 'auto' to use the
-     * previous geometry's value.
-     *
-     * It is not necessary to wrap this in (begin/end)Update() calls.
-     */
-    mx.mxCell.prototype.replaceGeometry = function (x, y, width, height, graph) {
-      const oldGeo = this.getGeometry();
-      const newGeo = new mx.mxGeometry(oldGeo.x, oldGeo.y, oldGeo.width, oldGeo.height);
-
-      if (x !== 'auto') {
-        newGeo.x = x;
-      }
-      if (y !== 'auto') {
-        newGeo.y = y;
-      }
-      if (width !== 'auto') {
-        newGeo.width = width;
-      }
-      if (height !== 'auto') {
-        newGeo.height = height;
-      }
-
-      // to save entries on undo stack, don't call setGeometry unless necessary
-      if (oldGeo.x === newGeo.x && oldGeo.y === newGeo.y
-        && oldGeo.width === newGeo.width && oldGeo.height === newGeo.height) {
-        return;
-      }
-
-      graph.getModel().setGeometry(this, newGeo);
-    };
-  }
-
-  /**
-   * Sets up rules for circuits' grouping/selection behavior
-   */
-  initGroupingRules() {
-    // Cells can be moved outside their parent's bounding box without being disowned
-    mx.mxGraphHandler.prototype.setRemoveCellsFromParent(false);
-    mx.mxGraph.prototype.setConstrainChildren(false);
-
-    /**
-     * Never act as though the backbone cell was clicked.
-     * If it was, act like the circuitContainer was clicked instead.
-     */
-    const defaultGetInitialCellForEvent = mx.mxGraphHandler.prototype.getInitialCellForEvent;
-    mx.mxGraphHandler.prototype.getInitialCellForEvent = function (me) {
-      let cell = defaultGetInitialCellForEvent.apply(this, arguments);
-      if (cell.isBackbone()) {
-        cell = cell.getCircuitContainer();
-      }
-      return cell;
-    }
-
-    /**
-     * For some reason, the above method doesn't work with alt-clicking.
-     * This method covers that case.
-     */
-    const defaultSelectCellForEvent = mx.mxGraph.prototype.selectCellForEvent;
-    mx.mxGraph.prototype.selectCellForEvent = function (cell, evt) {
-      if (cell.isBackbone())
-        cell = cell.getParent();
-
-      defaultSelectCellForEvent.apply(this, [cell, evt]);
-    };
-
-    /**
-     * Some methods of selecting cells don't involve clicking directly
-     * on the cell at all (for example rubberband selection).
-     * This to guarantees the backbone can never be selected, no matter what.
-     *
-     * (The previous two methods are still necessary, or clicking on the
-     * backbone would select nothing, instead of passing the click
-     * event up to the circuitContainer.)
-     */
-    mx.mxGraph.prototype.isCellSelectable = function (cell) {
-      return !cell.isBackbone();
-    }
-  }
-
-  /**
-   * Sets up all the constant styles used by the graph.
-   *
-   * Can only be called before this.graph is initialized
-   */
-  initStyles() {
-    // Main glyph settings. These are applied to sequence feature glyphs and molecular species glyphs
-    this.baseMolecularSpeciesGlyphStyle = {};
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_FILLCOLOR] = '#ffffff';
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_STROKECOLOR] = '#000000';
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_NOLABEL] = false;
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_VERTICAL_ALIGN] = 'top';
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_VERTICAL_LABEL_POSITION] = 'bottom';
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_EDITABLE] = false;
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_RESIZABLE] = 0;
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_DIRECTION] = "east";
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_STROKEWIDTH] = 2;
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_FONTCOLOR] = '#000000';
-    this.baseMolecularSpeciesGlyphStyle[mx.mxConstants.STYLE_FONTSIZE] = 14;
-    //this.baseGlyphStyle[mx.mxConstants.DEFAULT_HOTSPOT] = 0;
-
-    // Sequence features need almost the same styling as molecularSpecies
-    this.baseSequenceFeatureGlyphStyle = mx.mxUtils.clone(this.baseMolecularSpeciesGlyphStyle);
-    this.baseSequenceFeatureGlyphStyle[mx.mxConstants.STYLE_PORT_CONSTRAINT] = [mx.mxConstants.DIRECTION_NORTH, mx.mxConstants.DIRECTION_SOUTH];
-
-
-    const textBoxStyle = {};
-    textBoxStyle[mx.mxConstants.STYLE_SHAPE] = mx.mxConstants.SHAPE_LABEL;
-    textBoxStyle[mx.mxConstants.STYLE_FILLCOLOR] = '#ffffff';
-    textBoxStyle[mx.mxConstants.STYLE_STROKECOLOR] = '#000000';
-    textBoxStyle[mx.mxConstants.STYLE_FONTCOLOR] = '#000000';
-    this.graph.getStylesheet().putCellStyle(textboxStyleName, textBoxStyle);
-
-    const circuitContainerStyle = {};
-    circuitContainerStyle[mx.mxConstants.STYLE_SHAPE] = mx.mxConstants.SHAPE_RECTANGLE;
-    circuitContainerStyle[mx.mxConstants.STYLE_STROKECOLOR] = 'none';
-    circuitContainerStyle[mx.mxConstants.STYLE_FILLCOLOR] = 'none';
-    circuitContainerStyle[mx.mxConstants.STYLE_RESIZABLE] = 0;
-    circuitContainerStyle[mx.mxConstants.STYLE_EDITABLE] = false;
-    circuitContainerStyle[mx.mxConstants.STYLE_PORT_CONSTRAINT] = [mx.mxConstants.DIRECTION_NORTH, mx.mxConstants.DIRECTION_SOUTH];
-    this.graph.getStylesheet().putCellStyle(circuitContainerStyleName, circuitContainerStyle);
-
-    const backboneStyle = {};
-    backboneStyle[mx.mxConstants.STYLE_SHAPE] = mx.mxConstants.SHAPE_LINE;
-    backboneStyle[mx.mxConstants.STYLE_STROKECOLOR] = '#000000';
-    backboneStyle[mx.mxConstants.STYLE_STROKEWIDTH] = 2;
-    backboneStyle[mx.mxConstants.STYLE_RESIZABLE] = 0;
-    backboneStyle[mx.mxConstants.STYLE_EDITABLE] = false;
-    this.graph.getStylesheet().putCellStyle(backboneStyleName, backboneStyle);
-
-    // Interaction styles
-    const baseInteractionGlyphStyle = {};
-    baseInteractionGlyphStyle[mx.mxConstants.STYLE_STROKEWIDTH] = 2;
-    baseInteractionGlyphStyle[mx.mxConstants.STYLE_ENDSIZE] = 10;
-    baseInteractionGlyphStyle[mx.mxConstants.STYLE_STROKECOLOR] = '#000000';
-    baseInteractionGlyphStyle[mx.mxConstants.STYLE_FILLCOLOR] = '#000000';
-    baseInteractionGlyphStyle[mx.mxConstants.STYLE_EDITABLE] = false;
-    baseInteractionGlyphStyle[mx.mxConstants.STYLE_EDGE] = mx.mxConstants.EDGESTYLE_ORTHOGONAL;
-    baseInteractionGlyphStyle[mx.mxConstants.STYLE_ENDFILL] = 0;
-
-    const interactionControlStyle = mx.mxUtils.clone(baseInteractionGlyphStyle); // Inherit from the interaction defaults.
-    interactionControlStyle[mx.mxConstants.STYLE_ENDARROW] = mx.mxConstants.ARROW_DIAMOND;
-    this.graph.getStylesheet().putCellStyle(interactionGlyphBaseStyleName + interactionControlName, interactionControlStyle);
-
-    const interactionInhibitionStyle = mx.mxUtils.clone(baseInteractionGlyphStyle);
-    interactionInhibitionStyle[mx.mxConstants.STYLE_ENDARROW] = interactionInhibitionName;
-    interactionInhibitionStyle[mx.mxConstants.STYLE_ENDSIZE] = 15;
-    this.graph.getStylesheet().putCellStyle(interactionGlyphBaseStyleName + interactionInhibitionName, interactionInhibitionStyle);
-
-    const interactionStimulationStyle = mx.mxUtils.clone(baseInteractionGlyphStyle);
-    interactionStimulationStyle[mx.mxConstants.STYLE_ENDARROW] = mx.mxConstants.ARROW_BLOCK;
-    this.graph.getStylesheet().putCellStyle(interactionGlyphBaseStyleName + interactionStimulationName, interactionStimulationStyle);
-
-    const interactionProcessStyle = mx.mxUtils.clone(baseInteractionGlyphStyle);
-    interactionProcessStyle[mx.mxConstants.STYLE_ENDARROW] = mx.mxConstants.ARROW_BLOCK;
-    interactionProcessStyle[mx.mxConstants.STYLE_ENDFILL] = 1;
-    this.graph.getStylesheet().putCellStyle(interactionGlyphBaseStyleName + interactionProcessName, interactionProcessStyle);
-
-    const interactionDegradationStyle = mx.mxUtils.clone(baseInteractionGlyphStyle);
-    interactionDegradationStyle[mx.mxConstants.STYLE_ENDARROW] = interactionDegradationName;
-    interactionDegradationStyle[mx.mxConstants.STYLE_ENDSIZE] = 20;
-    this.graph.getStylesheet().putCellStyle(interactionGlyphBaseStyleName + interactionDegradationName, interactionDegradationStyle);
-  }
-
-  /**
-   * Loads glyph stencils and their names from the glyphService, and
-   * saves them to mxGraph's shape registry
-   */
-  initCustomShapes() {
-
-    let stencils = this.glyphService.getSequenceFeatureGlyphs();
-
-    for (const name in stencils) {
-      // Create a new copy of the stencil for the graph.
-      const stencil = stencils[name][0];
-      const centered = stencils[name][1];
-      let customStencil = new mx.mxStencil(stencil.desc); // Makes a deep copy
-
-      // Change the copied stencil for mxgraph
-      let origDrawShape = mx.mxStencil.prototype.drawShape;
-
-      if (centered) {
-        customStencil.drawShape = function (canvas, shape, x, y, w, h) {
-          h /= 2;
-          y += h / 2;
-          origDrawShape.apply(this, [canvas, shape, x, y, w, h]);
-        }
-      } else {
-        customStencil.drawShape = function (canvas, shape, x, y, w, h) {
-          h = h / 2;
-          origDrawShape.apply(this, [canvas, shape, x, y, w, h]);
-        }
-      }
-
-      // Add the stencil to the registry and set its style.
-      mx.mxStencilRegistry.addStencil(name, customStencil);
-
-      const newGlyphStyle = mx.mxUtils.clone(this.baseSequenceFeatureGlyphStyle);
-      newGlyphStyle[mx.mxConstants.STYLE_SHAPE] = name;
-      this.graph.getStylesheet().putCellStyle(sequenceFeatureGlyphBaseStyleName + name, newGlyphStyle);
-    }
-
-    // molecularSpecies glyphs are simpler, since we don't have to morph
-    // them to always be centred on the strand
-    stencils = this.glyphService.getMolecularSpeciesGlyphs();
-    for (const name in stencils) {
-      const stencil = stencils[name][0];
-      let customStencil = new mx.mxStencil(stencil.desc); // Makes of deep copy of the stencil.
-      mx.mxStencilRegistry.addStencil(name, customStencil);
-
-      const newGlyphStyle = mx.mxUtils.clone(this.baseMolecularSpeciesGlyphStyle);
-      newGlyphStyle[mx.mxConstants.STYLE_SHAPE] = name;
-      this.graph.getStylesheet().putCellStyle(molecularSpeciesGlyphBaseStyleName + name, newGlyphStyle);
-    }
-
-    // *** Define custom markers for edge endpoints ***
-
-    /**
-     * Returns a function that draws an Inhibition glyph
-     * @param endPoint The connection's endpoint (ie the coordinate of the anchor it's attached to)
-     * @param unitX The x part of a vector specifying the connection's direction
-     * @param unitY The y part of a vector specifying the connection's direction
-     * @param size The size of the connection head, directly from the connection's style
-     * @param source Boolean, true if this is the source endpoint, false if it is the terminal endpoint
-     * @param sw The stroke width
-     * @param filled Boolean, false if the connection head should have any transparency in the middle, true otherwise
-     */
-    let inhibitionMarkerDrawFunction = function (canvas, shape, type, endPoint, unitX, unitY, size, source, sw, filled) {
-      return function () {
-        canvas.begin();
-        canvas.moveTo(endPoint.x + (unitY * (size + sw) / 2), endPoint.y - (unitX * (size + sw) / 2));
-        canvas.lineTo(endPoint.x - (unitY * (size + sw) / 2), endPoint.y + (unitX * (size + sw) / 2));
-        canvas.stroke();
-      };
-    };
-    mx.mxMarker.addMarker(interactionInhibitionName, inhibitionMarkerDrawFunction);
-
-    /**
-     * Returns a function that draws a Process glyph
-     */
-    let degradationMarkerDrawFunction = function (canvas, shape, type, endPoint, unitX, unitY, size, source, sw, filled) {
-      const triangleTipX = endPoint.x - (unitX * (size + sw) / 2);
-      const triangleTipY = endPoint.y - (unitY * (size + sw) / 2);
-
-      const circleCenterX = endPoint.x - (unitX * size / 4);
-      const circleCenterY = endPoint.y - (unitY * size / 4);
-
-      const root2over2 = Math.sin(Math.PI / 4);
-
-      // Changing the parameter controls how far the connection's stem is drawn
-      endPoint.x = triangleTipX;
-      endPoint.y = triangleTipY;
-
-      return function () {
-
-        // CIRCLE
-        canvas.ellipse(circleCenterX - size / 4, circleCenterY - size / 4, size / 2, size / 2);
-        canvas.stroke();
-
-        // TRIANGLE
-        canvas.begin();
-        canvas.moveTo(triangleTipX, triangleTipY);
-        canvas.lineTo(triangleTipX - (unitY * size / 4) - (unitX * size / 2), triangleTipY + (unitX * size / 4) - (unitY * size / 2));
-        canvas.lineTo(triangleTipX + (unitY * size / 4) - (unitX * size / 2), triangleTipY - (unitX * size / 4) - (unitY * size / 2));
-        canvas.close();
-        canvas.fillAndStroke();
-
-        // SLASH (line through the circle)
-        canvas.begin();
-        canvas.moveTo(circleCenterX + root2over2 * size / 4, circleCenterY - root2over2 * size / 4);
-        canvas.lineTo(circleCenterX - root2over2 * size / 4, circleCenterY + root2over2 * size / 4);
-        canvas.stroke();
-        // This code makes the line rotate, but it makes more sense to always slash the same way
-        // canvas.moveTo(circleCenterX + (unitY * root2over2 * size / 4) + (unitX * root2over2 * size / 4),
-        //   circleCenterY - (unitX * root2over2 * size / 4) + (unitY * root2over2 * size / 4));
-        // canvas.lineTo(circleCenterX - (unitX * root2over2 * size / 4) - (unitY * root2over2 * size / 4),
-        //   circleCenterY + (unitX * root2over2 * size / 4) - (unitY * root2over2 * size / 4));
-      };
-    };
-    mx.mxMarker.addMarker(interactionDegradationName, degradationMarkerDrawFunction);
-
-    // label drawing
-    let graphService = this;
-    this.graph.convertValueToString = function (cell) {
-      if (cell.isSequenceFeatureGlyph() || cell.isMolecularSpeciesGlyph()) {
-        let info = graphService.getFromGlyphDict(cell.value);
-        if (!info) {
-          return cell.value;
-        } else if (info.name != null && info.name != '') {
-          return info.name;
-        } else {
-          return info.displayID;
-        }
-      } else if (cell.isCircuitContainer()) {
-        return '';
-      } else {
-        return cell.value;
-      }
-    }
-
-    // label truncation
-    this.graph.getLabel = function (cell) {
-      let label = this.convertValueToString(cell);
-      if (label) {
-        let geometry = graphService.graph.getModel().getGeometry(cell);
-        let fontSize = graphService.graph.getCellStyle(cell)[mx.mxConstants.STYLE_FONTSIZE];
-        let max = geometry.width / (fontSize * 0.7);
-
-        if (max < label.length) {
-          return label.substring(0, max) + '...';
-        }
-      }
-      return label;
-    }
-
-    let mxGraphViewGetPerimeterPoint = mx.mxGraphView.prototype.getPerimeterPoint;
-    mx.mxGraphView.prototype.getPerimeterPoint = function (terminal, next, orthogonal, border) {
-      let point = mxGraphViewGetPerimeterPoint.apply(this, arguments);
-      if (point != null) {
-        let perimeter = this.getPerimeterFunction(terminal);
-
-        if (terminal.text != null && terminal.text.boundingBox != null) {
-          let b = terminal.text.boundingBox.clone();
-          b.grow(3);
-
-          if (mx.mxUtils.rectangleIntersectsSegment(b, point, next)) {
-            point = perimeter(b, terminal, next, orthogonal);
-          }
-        }
-      }
-
-      return point;
-    }
-  }
-
-  /**
-   * Sets up logic for handling sequenceFeatureGlyph movement
-   */
-  initSequenceFeatureGlyphMovement() {
-    this.graph.addListener(mx.mxEvent.MOVE_CELLS, mx.mxUtils.bind(this, async function (sender, evt) {
-      // sender is the graph
-
-      sender.getModel().beginUpdate();
-      let cancelled = false;
-      try {
-        let movedCells = evt.getProperty("cells");
-        // important note: if a parent cell is moving, none of its children
-        // can appear here (even if they were also selected)
-
-        // sort cells: processing order is important
-        movedCells = movedCells.sort(function (cellA, cellB) {
-          if (cellA.getRootId() !== cellB.getRootId()) {
-            // cells are not related: choose arbitrary order (but still group by root)
-            return cellA.getRootId() < cellB.getRootId() ? -1 : 1;
-          } else {
-            // cells are in the same circuitContainer:
-            // must be in sequence order
-            let aIndex = cellA.getCircuitContainer(sender).getIndex(cellA);
-            let bIndex = cellB.getCircuitContainer(sender).getIndex(cellB);
-
-            return aIndex - bIndex;
-          }
-        });
-
-        // ownership change check
-        let containers = new Set<string>();
-        if (sender.getCurrentRoot()) {
-          let ownershipChange = false;
-          for (let cell of movedCells) {
-            if (cell.isSequenceFeatureGlyph()) {
-              ownershipChange = true;
-              containers.add(cell.getParent().getValue());
-            }
-          }
-          for (let container of Array.from(containers.values())) {
-            let glyphInfo;
-            if (sender.getCurrentRoot().isComponentView()) {
-              glyphInfo = this.getFromGlyphDict(sender.getCurrentRoot().getId());
-            } else {
-              glyphInfo = this.getFromGlyphDict(container);
-            }
-            if (ownershipChange && glyphInfo.uriPrefix != GlyphInfo.baseURI && !await this.promptMakeEditableCopy(glyphInfo.displayID)) {
-              cancelled = true;
-              // go check the finally block because I couldn't undo until after the end update
-              return;
-            }
-          }
-        }
-
-        // If two adjacent sequenceFeatureGlyphs were moved, they should be adjacent after the move.
-        // This loop finds all such sets of glyphs (relying on the sorted order) and sets them to
-        // have the same x position so there is no chance of outside glyphs sneaking in between
-        let streak;
-        for (let i = 0; i < movedCells.length; i += streak) {
-          streak = 1;
-          if (!movedCells[i].isSequenceFeatureGlyph()) {
-            continue;
-          }
-          // found a sequenceFeature glyph. A streak might be starting...
-          const baseX = movedCells[i].getGeometry().x;
-          const rootId = movedCells[i].getRootId();
-          let streakWidth = movedCells[i].getGeometry().width;
-
-          while (i + streak < movedCells.length
-            && movedCells[i + streak].isSequenceFeatureGlyph()
-            && rootId === movedCells[i + streak].getRootId()) {
-            let xToContinueStreak = baseX + streakWidth;
-            if (xToContinueStreak === movedCells[i + streak].getGeometry().x) {
-              // The next cell continues the streak
-              movedCells[i + streak].replaceGeometry(baseX, 'auto', 'auto', 'auto', sender);
-
-              streakWidth += movedCells[i + streak].getGeometry().width;
-              streak++;
-            } else {
-              // the next cell breaks the streak
-              break;
-            }
-          }
-        }
-
-        // now all sequence feature glyphs are sorted by x position in the order
-        // they should be when the move finishes.
-        // (equal x position means they should stay in the order they were previously in)
-        let cells = sender.getChildVertices(sender.getDefaultParent());
-        for (let circuitContainer of cells.filter(cell => cell.isCircuitContainer())) {
-          this.horizontalSortBasedOnPosition(circuitContainer);
-        }
-
-        // finallly, another special case: if a circuitContainer only has one sequenceFeatureGlyph,
-        // moving the glyph should move the circuitContainer
-        for (const cell of movedCells) {
-          if (cell.isSequenceFeatureGlyph() && cell.getParent().children.length === 2) {
-            const x = cell.getParent().getGeometry().x + evt.getProperty("dx");
-            const y = cell.getParent().getGeometry().y + evt.getProperty("dy");
-            cell.getParent().replaceGeometry(x, y, 'auto', 'auto', sender);
-          }
-        }
-
-        for (let container of Array.from(containers)) {
-          this.changeOwnership(container);
-        }
-
-        evt.consume();
-      } finally {
-        sender.getModel().endUpdate();
-        // undo has to happen after end update
-        if (cancelled) {
-          this.editor.undoManager.undo();
-          this.editor.undoManager.trim();
-        }
-      }
-    }));
-  }
 
   horizontalSortBasedOnPosition(circuitContainer) {
     // sort the children
